# docs/api/openapi_v1.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RFQListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RFQRead"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "maxLength": 256,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "RFQListResponse",
        "type": "object"
      },
...
      "RFQTextPreviewRequest": {
        "description": "Request body for the RFQ text preview endpoint.",
        "properties": {
          "channel_type": {
            "default": "BROKER_LME",
            "maxLength": 30,
            "title": "Channel Type",
            "type": "string"
          },
          "company_header": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Header"
          },
          "company_label_for_payoff": {
            "default": "Alcast",
            "maxLength": 100,
            "title": "Company Label For Payoff",
            "type": "string"
          },
          "leg1": {
            "$ref": "#/components/schemas/RFQLegInput"
          },
          "leg2": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RFQLegInput"
              },
              {
                "type": "null"
              }
            ]
          },
          "sync_ppt": {
            "default": false,
            "title": "Sync Ppt",
            "type": "boolean"
          },
          "trade_type": {
            "$ref": "#/components/schemas/RFQTradeTypeEnum"
          }
        },
        "required": [
          "trade_type",
          "leg1"
        ],
        "title": "RFQTextPreviewRequest",
        "type": "object"
      },
...
      "RFQTradeTypeEnum": {
        "enum": [
          "Swap",
          "Forward"
        ],
        "title": "RFQTradeTypeEnum",
        "type": "string"
      },
...
    }
  },
  "info": {
    "title": "Hedge Control Platform",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/rfqs": {
      "get": {
        "operationId": "list_rfqs_rfqs_get",
        "parameters": [
          {
            "description": "Filter by state (CREATED, SENT, QUOTED, AWARDED, CLOSED)",
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by state (CREATED, SENT, QUOTED, AWARDED, CLOSED)",
              "title": "State"
            }
          },
          {
            "description": "Filter by intent (COMMERCIAL_HEDGE, GLOBAL_POSITION, SPREAD)",
            "in": "query",
            "name": "intent",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by intent (COMMERCIAL_HEDGE, GLOBAL_POSITION, SPREAD)",
              "title": "Intent"
            }
          },
          {
            "description": "Filter by direction (BUY, SELL)",
            "in": "query",
            "name": "direction",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by direction (BUY, SELL)",
              "title": "Direction"
            }
          },
          {
            "description": "Filter by commodity",
            "in": "query",
            "name": "commodity",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by commodity",
              "title": "Commodity"
            }
          },
          {
            "description": "Include soft-deleted records",
            "in": "query",
            "name": "include_deleted",
            "required": false,
            "schema": {
              "default": false,
              "description": "Include soft-deleted records",
              "title": "Include Deleted",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RFQListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Rfqs",
        "tags": [
          "RFQs"
        ]
      },
      "post": {
        "operationId": "create_rfq_rfqs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RFQCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RFQRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Rfq",
        "tags": [
          "RFQs"
        ]
      }
    },
    "/rfqs/preview-text": {
      "post": {
        "description": "Generate RFQ message text without persisting anything.\n\nUseful for the trader to review the LME-formatted message before\nactually creating / sending the RFQ.",
        "operationId": "preview_rfq_text_rfqs_preview_text_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RFQTextPreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RFQTextPreviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preview Rfq Text",
        "tags": [
          "RFQs"
        ]
      }
    },
    "/rfqs/{rfq_id}": {
      "get": {
        "operationId": "get_rfq_rfqs__rfq_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "rfq_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Rfq Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RFQRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Rfq",
        "tags": [
          "RFQs"
        ]
      }
    },
    "/rfqs/{rfq_id}/actions/award": {
      "post": {
        "operationId": "award_rfq_rfqs__rfq_id__actions_award_post",
        "parameters": [
          {
            "in": "path",
            "name": "rfq_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Rfq Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RFQAwardRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RFQRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Award Rfq",
        "tags": [
          "RFQs"
        ]
      }
    },
    "/rfqs/{rfq_id}/actions/cancel": {
      "post": {
        "description": "Cancel an RFQ in CREATED or SENT state.",
        "operationId": "cancel_rfq_rfqs__rfq_id__actions_cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "rfq_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Rfq Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RFQCancelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RFQRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Rfq",
        "tags": [
          "RFQs"
        ]
      }
    },
    "/rfqs/{rfq_id}/actions/refresh": {
      "post": {
        "operationId": "refresh_rfq_rfqs__rfq_id__actions_refresh_post",
        "parameters": [
          {
            "in": "path",
            "name": "rfq_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Rfq Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RFQRefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RFQRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refresh Rfq",
        "tags": [
          "RFQs"
        ]
      }
    },
    "/rfqs/{rfq_id}/actions/refresh-counterparty": {
      "post": {
        "description": "Re-send invitation to a specific counterparty.",
        "operationId": "refresh_counterparty_rfqs__rfq_id__actions_refresh_counterparty_post",
        "parameters": [
          {
            "in": "path",
            "name": "rfq_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Rfq Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RFQRefreshCounterpartyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RFQRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refresh Counterparty",
        "tags": [
          "RFQs"
        ]
      }
    },
    "/rfqs/{rfq_id}/actions/reject": {
      "post": {
        "operationId": "reject_rfq_rfqs__rfq_id__actions_reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "rfq_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Rfq Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RFQRejectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RFQRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reject Rfq",
        "tags": [
          "RFQs"
        ]
      }
    },
    "/rfqs/{rfq_id}/actions/reject-quote": {
      "post": {
        "description": "Reject a specific counterparty quote without closing the RFQ.",
        "operationId": "reject_quote_rfqs__rfq_id__actions_reject_quote_post",
        "parameters": [
          {
            "in": "path",
            "name": "rfq_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Rfq Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "quote_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Quote Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RFQRejectQuoteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RFQRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reject Quote",
        "tags": [
          "RFQs"
        ]
      }
    },
    "/rfqs/{rfq_id}/archive": {
      "patch": {
        "operationId": "archive_rfq_rfqs__rfq_id__archive_patch",
        "parameters": [
          {
            "in": "path",
            "name": "rfq_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Rfq Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RFQUserActionBase"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RFQRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Archive Rfq",
        "tags": [
          "RFQs"
        ]
      }
    },
...
    "/rfqs/{rfq_id}/ranking": {
      "get": {
        "operationId": "get_spread_ranking_rfqs__rfq_id__ranking_get",
        "parameters": [
          {
            "in": "path",
            "name": "rfq_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Rfq Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpreadRankingRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Spread Ranking",
        "tags": [
          "RFQs"
        ]
      }
    },
...
  }
}